requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyodbc==5.2.0
qdrant-client==1.12.1
sentence-transformers==3.0.1
transformers==4.48.0
tokenizers>=0.21,<0.22
openai==1.58.1
google-genai==1.14.0
Pillow>=10.0,<12.0
python-multipart>=0.0.9,<0.0.21
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
apscheduler==3.10.4
redis>=5.0.0
numpy>=1.24,<2.0
huggingface-hub>=0.23,<0.26
pymongo>=4.10,<5.0
